reworked folder structure
