+++ client/src/App.tsx
@@ -1,13 +1,6 @@
 import './App.css';
-import {
-    BrowserRouter,
-    Navigate,
-    Route,
-    RouterProvider,
-    Routes,
-    createBrowserRouter,
-} from 'react-router-dom';
-import CampaignCreator from './components/CampaignCreator';
+import { BrowserRouter, Navigate, Route, RouterProvider, Routes, createBrowserRouter } from 'react-router-dom';
+import CampaignCreator from './components/campaignCreatorComponents/CampaignCreator';
 import DashboardRoutes from './components/DashboardRoutes';
 
 function App() {
@@ -15,14 +8,8 @@ function App() {
         <BrowserRouter>
             <div className="App">
                 <Routes>
-                    <Route
-                        path="/"
-                        element={<Navigate to="/twilight-library/dashboard" />}
-                    />
-                    <Route
-                        path="/twilight-library/dashboard/*"
-                        element={<DashboardRoutes />}
-                    />
+                    <Route path="/" element={<Navigate to="/twilight-library/dashboard" />} />
+                    <Route path="/twilight-library/dashboard/*" element={<DashboardRoutes />} />
                 </Routes>
             </div>
         </BrowserRouter>
