+++ client/src/components/Dashboard.tsx
@@ -1,9 +1,9 @@
-import PillarOptions from './PillarOptions';
-import OptionListCC from './OptionListCC';
+import PillarOptions from './campaignCreatorComponents/PillarOptions';
+import OptionListCC from './campaignCreatorComponents/OptionListCC';
 import ccData from '../static_data/campaign_creator.json';
-import RbListCC from './RbListCC';
+import RbListCC from './campaignCreatorComponents/RbListCC';
 import { Button, Heading, Flex, Separator } from '@radix-ui/themes';
-import { TLSelect } from './primitive_components/primitives';
+import { TLSelect } from './primitiveComponents/Primitives';
 import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
