+++ client/src/components/DashboardRoutes.tsx
@@ -1,15 +1,8 @@
-import {
-    BrowserRouter,
-    Navigate,
-    Route,
-    RouterProvider,
-    Routes,
-    createBrowserRouter,
-} from 'react-router-dom';
-import CampaignCreator from './CampaignCreator';
+import { BrowserRouter, Navigate, Route, RouterProvider, Routes, createBrowserRouter } from 'react-router-dom';
+import CampaignCreator from './campaignCreatorComponents/CampaignCreator';
 import Dashboard from './Dashboard';
-import TimelineCreator from './TimelineCreator';
-import CampaignCreatorRoutes from './CampaignCreatorRoutes';
+import TimelineCreator from './campaignCreatorComponents/TimelineCreator';
+import CampaignCreatorRoutes from './campaignCreatorComponents/CampaignCreatorRoutes';
 
 function DashboardRoutes() {
     return (
